Revert "refactor: split parts from plugin"
This reverts commit 53a1b8ea80d187e6ef59394fce4c98e2b73470d7.
