+++ packages/module-federation-metro/src/plugin.ts
@@ -5,18 +5,12 @@ import type { Resolution } from "metro-resolver";
 import generateManifest from "./generate-manifest";
 import { getModuleFederationSerializer } from "./serializer";
 import {
+  SharedConfig,
   ModuleFederationConfig,
   ModuleFederationConfigNormalized,
+  Shared,
 } from "./types";
 import { ConfigError } from "./utils/errors";
-import { createSharedModule, getSharedPath } from "./serializer/shared";
-import { createRemoteModule } from "./serializer/remote";
-import {
-  getRemoteEntryModule,
-  getRemoteHMRSetupModule,
-} from "./serializer/container";
-import { createRemoteModuleRegistryModule } from "./serializer/registry";
-import { createInitHostVirtualModule } from "./serializer/host";
 
 declare global {
   var __METRO_FEDERATION_CONFIG: ModuleFederationConfigNormalized;
@@ -32,10 +26,234 @@ const ASYNC_REQUIRE_REMOTE = "mf:async-require-remote";
 const MANIFEST_FILENAME = "mf-manifest.json";
 const DEFAULT_ENTRY_FILENAME = "remoteEntry.bundle";
 
+function getSharedString(options: ModuleFederationConfigNormalized) {
+  const shared = Object.keys(options.shared).reduce((acc, name) => {
+    acc[name] = `__SHARED_${name}__`;
+    return acc;
+  }, {} as Record<string, string>);
+
+  let sharedString = JSON.stringify(shared);
+  Object.keys(options.shared).forEach((name) => {
+    const sharedConfig = options.shared[name];
+    const entry = createSharedModuleEntry(name, sharedConfig);
+    sharedString = sharedString.replaceAll(`"__SHARED_${name}__"`, entry);
+  });
+
+  return sharedString;
+}
+
+function getEarlySharedDeps(shared: Shared) {
+  return Object.keys(shared).filter((name) => {
+    if (name === "react") return true;
+    if (name === "react-native") return true;
+    if (name.startsWith("react-native/")) return true;
+    return false;
+  });
+}
+
+function getInitHostModule(options: ModuleFederationConfigNormalized) {
+  const initHostPath = require.resolve("./runtime/init-host.js");
+  let initHostModule = fs.readFileSync(initHostPath, "utf-8");
+
+  const sharedString = getSharedString(options);
+
+  // must be loaded synchronously at all times
+  const earlySharedDeps = getEarlySharedDeps(options.shared);
+
+  // Replace placeholders with actual values
+  initHostModule = initHostModule
+    .replaceAll("__NAME__", JSON.stringify(options.name))
+    .replaceAll("__REMOTES__", generateRemotes(options.remotes))
+    .replaceAll("__SHARED__", sharedString)
+    .replaceAll("__EARLY_SHARED__", JSON.stringify(earlySharedDeps))
+    .replaceAll("__PLUGINS__", generateRuntimePlugins(options.plugins))
+    .replaceAll("__SHARE_STRATEGY__", JSON.stringify(options.shareStrategy));
+
+  return initHostModule;
+}
+
+function getRemoteModuleRegistryModule(
+  options: ModuleFederationConfigNormalized
+) {
+  const registryPath = require.resolve("./runtime/remote-module-registry.js");
+  let registryModule = fs.readFileSync(registryPath, "utf-8");
+
+  registryModule = registryModule.replaceAll(
+    "__EARLY_MODULE_TEST__",
+    "/^react(-native(\\/|$)|$)/"
+  );
+
+  return registryModule;
+}
+
+function createSharedModuleEntry(name: string, options: SharedConfig) {
+  const template = {
+    version: options.version,
+    scope: "default",
+    shareConfig: {
+      singleton: options.singleton,
+      eager: options.eager,
+      requiredVersion: options.requiredVersion,
+    },
+    get: options.eager
+      ? `__GET_SYNC_PLACEHOLDER__`
+      : `__GET_ASYNC_PLACEHOLDER__`,
+  };
+
+  const templateString = JSON.stringify(template);
+
+  return templateString
+    .replaceAll('"__GET_SYNC_PLACEHOLDER__"', `() => () => require("${name}")`)
+    .replaceAll(
+      '"__GET_ASYNC_PLACEHOLDER__"',
+      `async () => import("${name}").then((m) => () => m)`
+    );
+}
+
+function getRemoteModule(name: string) {
+  const remoteTemplatePath = require.resolve("./runtime/remote-module.js");
+
+  return fs
+    .readFileSync(remoteTemplatePath, "utf-8")
+    .replaceAll("__MODULE_ID__", `"${name}"`);
+}
+
 function createMFRuntimeNodeModules(projectNodeModulesPath: string) {
   const mfMetroPath = path.join(projectNodeModulesPath, ".mf-metro");
   fs.mkdirSync(mfMetroPath, { recursive: true });
   return mfMetroPath;
+}
+
+function generateRuntimePlugins(runtimePlugins: string[]) {
+  const pluginNames: string[] = [];
+  const pluginImports: string[] = [];
+
+  runtimePlugins.forEach((plugin, index) => {
+    const pluginName = `plugin${index}`;
+    pluginNames.push(`${pluginName}()`);
+    pluginImports.push(`import ${pluginName} from "${plugin}";`);
+  });
+
+  const imports = pluginImports.join("\n");
+  const plugins = `const plugins = [${pluginNames.join(", ")}];`;
+
+  return `${imports}\n${plugins}`;
+}
+
+function generateRemotes(remotes: Record<string, string> = {}) {
+  const remotesEntries: string[] = [];
+  Object.entries(remotes).forEach(([remoteAlias, remoteEntry]) => {
+    const remoteEntryParts = remoteEntry.split("@");
+    const remoteName = remoteEntryParts[0];
+    const remoteEntryUrl = remoteEntryParts.slice(1).join("@");
+
+    remotesEntries.push(
+      `{ 
+          alias: "${remoteAlias}", 
+          name: "${remoteName}", 
+          entry: "${remoteEntryUrl}", 
+          entryGlobalName: "${remoteName}", 
+          type: "var" 
+       }`
+    );
+  });
+
+  return `[${remotesEntries.join(",\n")}]`;
+}
+
+function getRemoteEntryModule(options: ModuleFederationConfigNormalized) {
+  const remoteEntryTemplatePath = require.resolve("./runtime/remote-entry.js");
+  let remoteEntryModule = fs.readFileSync(remoteEntryTemplatePath, "utf-8");
+
+  const sharedString = getSharedString(options);
+  const earlySharedDeps = getEarlySharedDeps(options.shared);
+
+  const exposes = options.exposes || {};
+
+  const exposesString = Object.keys(exposes)
+    .map((key) => {
+      const importName = path.relative(".", exposes[key]);
+      const importPath = `../../${importName}`;
+
+      return `"${key}": async () => {
+          const module = await import("${importPath}");
+          return module;
+        }`;
+    })
+    .join(",");
+
+  return remoteEntryModule
+    .replaceAll("__PLUGINS__", generateRuntimePlugins(options.plugins))
+    .replaceAll("__SHARED__", sharedString)
+    .replaceAll("__REMOTES__", generateRemotes(options.remotes))
+    .replaceAll("__EARLY_SHARED__", JSON.stringify(earlySharedDeps))
+    .replaceAll("__EXPOSES_MAP__", `{${exposesString}}`)
+    .replaceAll("__NAME__", `"${options.name}"`)
+    .replaceAll("__SHARE_STRATEGY__", JSON.stringify(options.shareStrategy));
+}
+
+function getRemoteHMRSetupModule() {
+  const remoteHMRSetupTemplatePath = require.resolve("./runtime/remote-hmr.js");
+  let remoteHMRSetupModule = fs.readFileSync(
+    remoteHMRSetupTemplatePath,
+    "utf-8"
+  );
+
+  return remoteHMRSetupModule;
+}
+
+function createInitHostVirtualModule(
+  options: ModuleFederationConfigNormalized,
+  vmDirPath: string
+) {
+  const initHostModule = getInitHostModule(options);
+  const initHostPath = path.join(vmDirPath, "init-host.js");
+  fs.writeFileSync(initHostPath, initHostModule, "utf-8");
+  return initHostPath;
+}
+
+// virtual module: remote-module-registry
+function createRemoteModuleRegistryModule(
+  options: ModuleFederationConfigNormalized,
+  vmDirPath: string
+) {
+  const registryModule = getRemoteModuleRegistryModule(options);
+  const registryPath = path.join(vmDirPath, "remote-module-registry.js");
+  fs.writeFileSync(registryPath, registryModule, "utf-8");
+  return registryPath;
+}
+
+function createSharedModule(sharedName: string, outputDir: string) {
+  const sharedFilePath = getSharedPath(sharedName, outputDir);
+  // we need to create the shared module if it doesn't exist
+  const sharedModule = getRemoteModule(sharedName);
+  fs.mkdirSync(path.dirname(sharedFilePath), { recursive: true });
+  fs.writeFileSync(sharedFilePath, sharedModule, "utf-8");
+  return sharedFilePath;
+}
+
+function getSharedPath(name: string, dir: string) {
+  const sharedName = name.replaceAll("/", "_");
+  const sharedDir = path.join(dir, "shared");
+  return path.join(sharedDir, `${sharedName}.js`);
+}
+
+function createRemoteModule(name: string, outputDir: string) {
+  const remoteModule = getRemoteModule(name);
+  const remoteFilePath = getRemoteModulePath(name, outputDir);
+  fs.mkdirSync(path.dirname(remoteFilePath), { recursive: true });
+  fs.writeFileSync(remoteFilePath, remoteModule, "utf-8");
+  return remoteFilePath;
+}
+
+function getRemoteModulePath(name: string, outputDir: string) {
+  const remoteModuleName = name.replaceAll("/", "_");
+  const remoteModulePath = path.join(
+    outputDir,
+    "remote",
+    `${remoteModuleName}.js`
+  );
+  return remoteModulePath;
 }
 
 function stubSharedModules(
