--- packages/module-federation-metro/src/serializer/plugins.ts
@@ -1,15 +1,0 @@
-export function generateRuntimePlugins(runtimePlugins: string[]) {
-  const pluginNames: string[] = [];
-  const pluginImports: string[] = [];
-
-  runtimePlugins.forEach((plugin, index) => {
-    const pluginName = `plugin${index}`;
-    pluginNames.push(`${pluginName}()`);
-    pluginImports.push(`import ${pluginName} from "${plugin}";`);
-  });
-
-  const imports = pluginImports.join("\n");
-  const plugins = `const plugins = [${pluginNames.join(", ")}];`;
-
-  return `${imports}\n${plugins}`;
-}
